sanitize auto build path in sample text output
It was brought to my attention that the build process results in
committing the project path of the person who built the site.

I chose to use /Users/jim/Downloads/git_tutorial/work

/Users/jim because the rest of the labs are written from the perspective
of Jim.

Downloads/git_tutorial/ because Jim used a mac and that's the location
the zip file would likely end up after downloading the zip file from the
website.

And finally, the work directory is the directory that the labs expect
the user to create.